+++ FinalPJT/app/src/main/java/com/ssafy/finalpjt/database/repository/TodoRepository.kt
@@ -14,7 +14,7 @@ class TodoRepository (private val db: CarrotDatabase){
 
     private val todoDao=db.todoDao()
 
-    fun getTodayTodo(date:Int):LiveData<MutableList<Todo>>{
+    fun getTodayTodo(date:Long):LiveData<MutableList<Todo>>{
         return todoDao.getTodayTodo(date)
     }
     suspend fun insertTodo(todo: Todo)=db.withTransaction{
